requirements.txt
# FastAPI framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# File upload handling
python-multipart==0.0.9

# Image processing
pillow==10.2.0

# Ollama client
ollama==0.1.6

# Data validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Rate limiting
slowapi==0.1.9

# Optional: Redis for caching (comentar se não usar)
redis==5.0.1

# Logging and monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.25.2

# Streamlit (manter seu app atual)
streamlit==1.31.0
